Remove hardcoded Azure region when creating machinset. (#2991)
Recently, the Azure regions have been modified per release to allow
balancing the number of deployments per region [0]. However, the
getMachineset function was using a harcoded centralus region when
creating a machineset on Azure. This harcoded region was making
4.10 (which by default were changed into westus region) fail and
the same will occur for 4.11.

This commit retrieves the Azure region from one of the machines
created by the flexy-templates and replaces it in the machineset
manifest used by the winc tests.

Also, removes the harcoded usage of the machineset file into using
the machineset file provided by the getMachineset function. Otherwise,
we would be creating a different machineset than what we pass when
invoking the function.

[0] - https://gitlab.cee.redhat.com/aosqe/flexy-templates/-/commit/acdc970c15dadfc44a468126d4cad63d4797e39a