Add support for vsphere bastion in winc tests. (#3091)
Up to know, all test cases which required the bastion host to
run a command in some of the Windows machine workers was failing
in vsphere.

This patch adds support for vsphere, as well as it filters out
the ipv6 internal addresses (as we want to ssh only in the ipv4
ones) in the fetchAddress utils function.

Also, cleaning up unneeded parameters in function
checkRunningServicesOnWindowsNode.