Remove port-forward from winc 33768 test. (#3152)
Test case 33768 performs a query to the prometheus service by
running a port-forward of the 9090 prometheus' service. However,
this does not work correctly in vSphere clusters as there is
an SSH bastion in between, so oc port-forward does not work as
expected.

This commit modifies the test case so that we run the curl query
directly into the pod. This way, there is no need for port
forwarding and the test will be able to run in all platforms.