[OCPQE-10892] Removing host_service_logs/windows/log_winevent/docker_winevent.log from logs to be checked in must_gather (#3889)
In one of our tests, the one which checks the consistency of the built logs during must-gather, the log host_service_logs/windows/log_winevent/docker_winevent.log is also checked to be collected. Our question is, do we need to keep that log? I mean, what would happen if we are running the test in a worker which does not even have docker installed? this log would not be present, right?

Developers' PR to solve this: https://github.com/openshift/must-gather/pull/315