[OCPQE-9170] Fix waitUntilWMCOStatusChanged functionality. (#3922)
Currently, this function was printing out the whole log (whic can
take up to several MB) to check if a new log had made appearence.
This patch solves that by printing out only the last 10 seconds
of the log everytime the log is retrieved.

Also, it fixes the logic to stop the wait once the string we
are searching for is found in the log.