Fix and simplify the multi-os test case OCP-37096. (#4393)
This patch depends on [0] to work properly, as we read
the primary and secondary's operating system information from
the config-map created in flexy-templates.

This modification allows us creating a second machineset with
a different OS than the one use for the machinesets during deployment.
For example, if we deploy Windows workers on Windows 2019, the
secondary OS will be Windows 2022 and the other way around.

This commit also modifies the behavior of createWindowsWorkload to
allow modifying multiple tags in the workload's manifest, not only
one as it was done up to now.

[0] - https://gitlab.cee.redhat.com/aosqe/flexy-templates/-/merge_requests/1386