Increase waiting timeout for Windows machines and workloads. (#4522)
It has been observed in some cloud providers, that 10 minutes
isn't enough time to wait for the machines to be provisioned.
However, after few tests, 15 seems like a reasonable time and
the tests are more stable.

This patch increases the timeout from 5 minutes to 10 for the
workloads in scaleDeployment function. It is not consistent that
we wait 15 minutes when creating some workload, but only 5 when
scaling up.