[OCPQE-8991] Automating OCP-35707 (#5821)
Automation of OCP-35707, which simulates an upgrade/update
operation by modifying the version annotation. In parallel, a check
of the LoadBalancer service (via curl on the LB address) using a
goroutine is performed to make sure there is no impact on the services
running.

This mechanism to check LB connectivity without blocking has been added
in test case 38186.

Co-authored-by: Ronnie Rasouli <rrasouli@redhat.com>