Adapt WINC test cases to run properly on GCP (#6272)
Some of the specifics for GCP required adaptation in the test cases,
like including new templates for MachineSets compatible with GCP or new
required labels to spawn a Windows node without WMCO processing it.

Also, it improves the getWMCOVersionFromLogs
function to use regexp submatch, as there were several matches for
the version field in the logs.

This patch fixes some indentation issues added to the autoscaler
template as well as adds the label -l=app=win-webserver when checking
containers running on the windows nodes.