Refactor winc (#6301)
* Fix get machineset in utils

The way we fetched our machineset was made by oc command.
We used to get the machienset from the current cluster assuming that there is only 1 machineset, in case there are several machinesets then we get a list of items which can fail that function.
This change is also a refactoring to our code, by using more constant variables, cleaner code.

* Fix autoscaler template indentation problem, remove sleep in 42496