Fix autoscaler on vsphere. (#7359)
In [0], the code added broke the autoscaller test for vsphere as it relies on the winc
machineset. So, as the code is up to now, the getWindowsMachineSetName would return
winworker as MachineSet name even though the MachineSet we are creating is named winc.
This caused the existing winworker MachineSet to even be removed after the test failed,
causing the tests comming after it to fail.
Also, as part of the tune in, the number of workloads needed to trigger the autoscaler to
scale up the MachineSet seems to be 4, not 2.