Add Terraform library and credentials function for vSphere and GCP. (#6881)
This PR is an attempt to split up #5271 which grew consistently
up to a point in which the review is unmanageable. This first PR
submits the common pieces of code, the inclusion of the terraform.go
library, as well as the addition of the helper functions which
retrieve the cloud provider credentials (AWS and Azure already
existed) for GCP and vSphere. These credentials are required by
Terraform [0][1] to be able to provision the infrastructure resources
into the cloud.

This patch also adds a set of unit tests for the Terraform library.

[0] - https://registry.terraform.io/providers/hashicorp/google/latest/docs/guides/provider_reference
[1] - https://registry.terraform.io/providers/hashicorp/vsphere/latest/docs#argument-reference