Automate OCP-37086: WMCO in custom namespace (#7553)
This patch adds a new testcase, OCP 37086, which verifies the
installation of WMCO in a different namespace than
openshift-windows-machine-config-operator.

As part of the changes required for the test, the logic in setConfigmap,
as well as its name is being changed into createManifestFile, a generic
function which allows us creating manifes files (oc create -f ) passing
arguments for replacing.