Add test coverage for kube-log-runner. (#7660)
In [0] support for the kube-log-runner binary was added, this involves
having a new binary in the payload/kube directory named kube-log-runner.
Also, the log which was printed out before by the kubelet and kube-proxy
services: 'Log file created at' is not present anymore. Therefore, the
logic has been changed to obtain the kubelet and kube-proxy logs for each
of the nodes and try to grep the appearance of 'Running <service> as a
Windows service!'.

[0] - https://issues.redhat.com/browse/WINC-962