Support for parallel execution of Terraform library. (#7887)
One of the impediments of the terraform.go utils library was
that when invoking InstallTerraform it was always downloading
the terraform binary. This patch adds logic to identify
programmatically which binary to use and if no options are
available, then it will be downloaded in /tmp/terraform_*

At the same time, this patch deletes the RemoveTerraform
function as we don't really have to get rid of the binary
anymore if we want to use the binary installed in PATH, for
example.