reverting #7526 since OCPBUGS-3783 is now fixed (#8908)
previously, due to [OCPBUGS-3783](https://issues.redhat.com/browse/OCPBUGS-3783) we had set test case OCP-56072 as flaky.  

since the bug has been fixed in 4.14, and afaik this is where master is pointing currently, submitting the revert as this separate PR (for 4.14 only at this stage), while a follow up cherry pick is expected once the bugfix is backpacked further.