[OCPQE-14983] Remove check for order at which services stopped. (#9049)
Due to the amount of failures the test case Author:jfrancoa-Medium-56354-Stop
dependent services before stopping a service in WICD [Disruptive] is having.
And taking into account that Developer's feedback was that the order at which
the services were logged wasn't really relevant. It is better to just remove
the check to add stability to our test base.