[OCPQE-15173] Automate OCP-66352. (#11462)
This test adds a new storage.go file in which storage test cases
will be added. The test case takes care of installing the CSI driver
for the Windows nodes and configure the persistent storage for the
workloads created, using that CSI driver. So far only vSphere has
been implemented.

It also modifies the existing test cases to take this new binary
csi-proxy into account.