Fix missing parenthesis in logical statement. (#12816)
The code added for 4.13 to handled Nutanix broke vSphere as a
parenthesis was missing to englobe the condition that, if iaasPlatform
is either vSphere or Nutanix AND the name is windows then enter
the if condition. Without the parenthesis, as it was, it will consider
each logical statement separetely, so as soon as iaasPlatform is
vSphere it will enter the if statement (which isn't what we really want).
Adding the parenthesis fixes the problem.