Fix getNodeNameFromIP in winc.go. (#12829)
Before, the getNodeNameFromIP was relying on a hardcoded index in which
we would expect the InternalIP field to appear from all the items in the
status.addresses. However, this approach is error prone, as the structure
is not guaranteed, and sometimes the InternalIP item was in position 0, others
in 1, other times in 2.

This patch fixes the error by using a go-template instead, which will iterate
over all nodes in the json retured by oc get nodes command and for each node,
it will iterate over each status.addresses item. If the .address field for
that item matches the IP passed as argument, then we return as input the
node .metadata.name field.

It also increases the timeout for the InternalIP address to appear,
which seems to be taking more than 200 seconds on vSphere.