OCPQE-15773: Pivot to the new RDU2 Bare Metal Infrastructure and Fix UPI BM @RDU2 client and functions (#18967)
The new infrastructure does not expose the BMC/OOB network and has not anymore a web service to provide the information previously consumed by the cases related to this code:
- IPMI is now reachable through port forwarding.
- The hosts.yaml file available in the SHARED_DIR will provide the required information about the nodes.

Finally, fixes improve the quality of this code, where possible and having minimal impact on the other code.

Refers to gebn/bmc#46 with regard to the privilege escalation to power on/off the nodes.