Add function to set and unset proxy to reach bmc (#21711)
To verify that vmedia is attached to the host, we need to query redfish
API of the bmc address that is only accessible via proxy. The function
checks if proxy variables are already set before setting or unsetting
them.

Removed 'Serial' from 'Diruptive' tests as Disruptive implies Serial and
to keep it consistent and avoid confusion in future